Fix cleanup condition in on_build_error to check if temp_dir exists

+++ mkdocs_multirepo_plugin/plugin.py
@@ -358,7 +358,7 @@ class MultirepoPlugin(BasePlugin):
             shutil.rmtree(str(self.temp_dir))
 
     def on_build_error(self, error):
-        if self.temp_dir and self.config.get("cleanup"):
+        if self.temp_dir and self.config.get("cleanup") and self.temp_dir.exists():
             temp_dir = self.config.get("temp_dir")
             log.info(f"Multirepo plugin is cleaning up {temp_dir}/")
             shutil.rmtree(str(self.temp_dir))
